LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-shadow-container-expected.txt


PASS Can query named container in shadow
PASS ::part() can not see inner named containers
FAIL ::slotted() can see inner named containers assert_equals: expected "10px 5px" but got "10px 10px"
